+++ R/class_experiment.R
@@ -11,6 +11,7 @@
 #' \item{\code{.model}:}{Internal. The model associated with the iteration.}
 #' \item{\code{.group}:}{Internal. The group associated with the iteration.}
 #' \item{\code{.iter}:}{Internal. The iteration number.}
+#' \item{\code{.seed}:}{The seed used to generate the experiment.}
 #' }
 #' @rdname CalmrExperiment
 #' @exportClass CalmrExperiment
@@ -28,7 +29,8 @@ methods::setClass(
     results = "CalmrExperimentResult",
     .model = "character",
     .group = "character",
-    .iter = "integer"
+    .iter = "integer",
+    .seed = "ANY"
   )
 )