+++ NEWS.md
@@ -1,4 +1,6 @@
 # calmr 0.6.3
+* Corrected some issues in directional models.
+* Created a vignette to expose the behaviour of directional models.
 * Removed randomization column requirement from designs. Randomization of phases is now specified using the "!" character anywhere in the phase string.
 * Added support for seed experiment generation in `make_experiment()`.
 * Added `set_calmr_palette()` function to control the colour/fill scales used to plot results ([#1](https://github.com/victor-navarro/calmr/issues/1)).